requirements.txt
fusionauth-client>=1.36.0
itsdangerous>=2.1.2
Jinja2>=3.1.2
pkce>=1.0.3
starlette>=0.21.0
uvicorn>=0.18.3
